Adds home and about page routes - create a basic structure to page components.
