+++ src/App.js
@@ -1,28 +1,27 @@
-import { MoviesContainer } from './styles/MovieContainer';
 import { Header } from './components/Header';
-import { MoviesList } from './components/MoviesList';
-import { useEffect, useState } from 'react';
-import { MoviesService } from './api/MoviesService';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+} from 'react-router-dom';
+import { About } from './views/About';
+import { Home } from './views/Home';
 
 function App() {
-  const [movies, setMovies] = useState([]);
-
-  const fetchMovies = async () => {
-    const { data: { results = [] } = {} } = await MoviesService.getMovies();
-    setMovies(results);
-  }
-
-  useEffect(() => {
-    fetchMovies();
-  }, []);
-
   return (
     <div className="App">
-      <Header />
+      <Router>
+        <Header />
 
-      <MoviesContainer>
-        <MoviesList movies={movies} />
-      </MoviesContainer>
+        <Switch>
+          <Route path="/about">
+            <About />
+          </Route>
+          <Route path="/">
+            <Home />
+          </Route>
+        </Switch>
+      </Router>
     </div>
   );
 }
