+++ src/views/About.js
@@ -1,0 +1,1 @@
+export const About = () => <h1>Página Sobre</h1>
